requirements.txt
streamlit
pandas
ortools
reportlab
openpyxl
xlsxwriter
matplotlib
python-dotenv